// src/client/src/store/modals/typeKeys.ts
export enum MODAL_TYPES {
  AZURE_LOGIN_MODAL = "AZURE_LOGIN_MODAL",
  COSMOS_DB_MODAL = "COSMOS_DB_MODAL",
  GEN_MODAL = "GEN_MODAL",
  PRIVACY_MODAL = "PRIVACY_MODAL",
  VIEW_LICENSES_MODAL = "VIEW_LICENSES_MODAL",
  APP_SERVICE_MODAL = "APP_SERVICE_MODAL",
  ADD_PAGES_MODAL = "ADD_PAGES_MODAL"
}

export enum MODAL_TYPEKEYS {
  OPEN_MODAL = "WTS/modals/OPEN_MODAL",
  CLOSE_MODALS = "WTS/modals/CLOSE_MODALS"
}

export type ModalType =
  | null
  | MODAL_TYPES.COSMOS_DB_MODAL
  | MODAL_TYPES.GEN_MODAL
  | MODAL_TYPES.PRIVACY_MODAL
  | MODAL_TYPES.VIEW_LICENSES_MODAL
  | MODAL_TYPES.APP_SERVICE_MODAL
  | MODAL_TYPES.ADD_PAGES_MODAL
  | MODAL_TYPES.AZURE_LOGIN_MODAL

export interface ModalState {
  modalType: ModalType;
  modalData: any;
}